Page_Stock

+++ Project_Gestion_Stock/src/main/java/com/gs/project_gestion_stock/service/StockService.java
@@ -21,7 +21,7 @@ public class StockService {
         return stockRepository.findAll().stream().map(stock -> {
             StockDTO dto = new StockDTO();
             dto.setId(stock.getId());
-            dto.setQuantiteDisponible(stock.getQuantite_disponible());
+            dto.setQuantiteImporte(stock.getQuantite_importe());
             dto.setQuantiteReserver(stock.getQuantite_reserver());
             dto.setSeuilReapprovisionnement(stock.getSeuil_reapprovisionnement());
             dto.setProduitNom(stock.getProduit() != null ? stock.getProduit().getNom() : null);
@@ -32,23 +32,31 @@ public class StockService {
     public Optional<Stock> getStockByID(int id){
         return stockRepository.findById(id);
     }
+
+
     public Stock createStock(Stock stock){
         return stockRepository.save(stock);
     }
+
+
     public Stock updateStock(int id, Stock newStock) {
         return stockRepository.findById(id).map(stock -> {
-            stock.setQuantite_disponible(newStock.getQuantite_disponible());
+            stock.setQuantite_importe(newStock.getQuantite_importe());
             stock.setQuantite_reserver(newStock.getQuantite_reserver());
             stock.setSeuil_reapprovisionnement(newStock.getSeuil_reapprovisionnement());
+            stock.setDate_expiration(newStock.getDate_expiration());
             stock.setProduit(newStock.getProduit());
             stock.setEmplacement(newStock.getEmplacement());
             return stockRepository.save(stock);
-        }).orElseGet(() -> {
-            newStock.setId(id);
-            return stockRepository.save(newStock);
-        });
+        }).orElseThrow(() -> new RuntimeException("Stock non trouvé avec l'ID : " + id));
     }
+
     public void deleteStock(int id) {
+        if (!stockRepository.existsById(id)) {
+            throw new RuntimeException("Le stock avec l'ID " + id + " n'existe pas.");
+        }
         stockRepository.deleteById(id);
     }
+
+
 }
